UI: split cron view modules
