+++ ui/src/ui/views/cron/filters.ts
@@ -1,0 +1,64 @@
+import { html } from "lit";
+import { t } from "../../../i18n/index.ts";
+import type { CronDeliveryStatus, CronRunsStatusValue } from "../../types.ts";
+
+export function getRunStatusOptions(): Array<{ value: CronRunsStatusValue; label: string }> {
+  return [
+    { value: "ok", label: t("cron.runs.runStatusOk") },
+    { value: "error", label: t("cron.runs.runStatusError") },
+    { value: "skipped", label: t("cron.runs.runStatusSkipped") },
+  ];
+}
+
+export function getRunDeliveryOptions(): Array<{ value: CronDeliveryStatus; label: string }> {
+  return [
+    { value: "delivered", label: t("cron.runs.deliveryDelivered") },
+    { value: "not-delivered", label: t("cron.runs.deliveryNotDelivered") },
+    { value: "unknown", label: t("cron.runs.deliveryUnknown") },
+    { value: "not-requested", label: t("cron.runs.deliveryNotRequested") },
+  ];
+}
+
+export function renderRunFilterDropdown(params: {
+  id: string;
+  title: string;
+  summary: string;
+  options: Array<{ value: string; label: string }>;
+  selected: string[];
+  onToggle: (value: string, checked: boolean) => void;
+  onClear: () => void;
+}) {
+  return html`
+    <div class="field cron-filter-dropdown" data-filter=${params.id}>
+      <span>${params.title}</span>
+      <details class="cron-filter-dropdown__details">
+        <summary class="btn cron-filter-dropdown__trigger">
+          <span>${params.summary}</span>
+        </summary>
+        <div class="cron-filter-dropdown__panel">
+          <div class="cron-filter-dropdown__list">
+            ${params.options.map(
+              (option) => html`
+                <label class="cron-filter-dropdown__option">
+                  <input
+                    type="checkbox"
+                    value=${option.value}
+                    .checked=${params.selected.includes(option.value)}
+                    @change=${(event: Event) => {
+                      const target = event.target as HTMLInputElement;
+                      params.onToggle(option.value, target.checked);
+                    }}
+                  />
+                  <span>${option.label}</span>
+                </label>
+              `,
+            )}
+          </div>
+          <div class="row">
+            <button class="btn" type="button" @click=${params.onClear}>${t("cron.runs.clear")}</button>
+          </div>
+        </div>
+      </details>
+    </div>
+  `;
+}
