+++ ui/src/ui/views/cron/helpers.ts
@@ -1,0 +1,45 @@
+import { html, nothing } from "lit";
+import { t } from "../../../i18n/index.ts";
+import { formatRelativeTimestamp } from "../../format.ts";
+
+export function toggleSelection<T extends string>(selected: T[], value: T, checked: boolean): T[] {
+  const set = new Set(selected);
+  if (checked) {
+    set.add(value);
+  } else {
+    set.delete(value);
+  }
+  return Array.from(set);
+}
+
+export function summarizeSelection(selectedLabels: string[], allLabel: string) {
+  if (selectedLabels.length === 0) {
+    return allLabel;
+  }
+  if (selectedLabels.length <= 2) {
+    return selectedLabels.join(", ");
+  }
+  return `${selectedLabels[0]} +${selectedLabels.length - 1}`;
+}
+
+export function renderSuggestionList(id: string, options: string[]) {
+  const clean = Array.from(new Set(options.map((option) => option.trim()).filter(Boolean)));
+  if (clean.length === 0) {
+    return nothing;
+  }
+  return html`<datalist id=${id}>
+    ${clean.map((value) => html`<option value=${value}></option> `)}
+  </datalist>`;
+}
+
+export function formatStateRelative(ms?: number) {
+  if (typeof ms !== "number" || !Number.isFinite(ms)) {
+    return t("common.na");
+  }
+  return formatRelativeTimestamp(ms);
+}
+
+export function formatRunNextLabel(nextRunAtMs: number, nowMs = Date.now()) {
+  const rel = formatRelativeTimestamp(nextRunAtMs);
+  return nextRunAtMs > nowMs ? t("cron.runEntry.next", { rel }) : t("cron.runEntry.due", { rel });
+}
