+++ ui/src/ui/views/cron/jobs.ts
@@ -1,0 +1,214 @@
+import { html, nothing } from "lit";
+import { t } from "../../../i18n/index.ts";
+import { formatMs } from "../../format.ts";
+import { formatCronSchedule } from "../../presenter.ts";
+import type { CronJob } from "../../types.ts";
+import type { CronProps } from "../cron.ts";
+import { formatStateRelative } from "./helpers.ts";
+
+export function buildChannelOptions(props: CronProps): string[] {
+  const options = ["last", ...props.channels.filter(Boolean)];
+  const current = props.form.deliveryChannel?.trim();
+  if (current && !options.includes(current)) {
+    options.push(current);
+  }
+  const seen = new Set<string>();
+  return options.filter((value) => {
+    if (seen.has(value)) {
+      return false;
+    }
+    seen.add(value);
+    return true;
+  });
+}
+
+export function resolveChannelLabel(props: CronProps, channel: string): string {
+  if (channel === "last") {
+    return "last";
+  }
+  const meta = props.channelMeta?.find((entry) => entry.id === channel);
+  if (meta?.label) {
+    return meta.label;
+  }
+  return props.channelLabels?.[channel] ?? channel;
+}
+
+export function renderJob(job: CronJob, props: CronProps) {
+  const isSelected = props.runsJobId === job.id;
+  const itemClass = `list-item list-item-clickable cron-job${isSelected ? " list-item-selected" : ""}`;
+  const selectAnd = (action: () => void) => {
+    props.onLoadRuns(job.id);
+    action();
+  };
+  return html`
+    <div class=${itemClass} @click=${() => props.onLoadRuns(job.id)}>
+      <div class="list-main">
+        <div class="list-title">${job.name}</div>
+        <div class="list-sub">${formatCronSchedule(job)}</div>
+        ${renderJobPayload(job)}
+        ${job.agentId ? html`<div class="muted cron-job-agent">${t("cron.jobDetail.agent")}: ${job.agentId}</div>` : nothing}
+      </div>
+      <div class="list-meta">
+        ${renderJobState(job)}
+      </div>
+      <div class="cron-job-footer">
+        <div class="chip-row cron-job-chips">
+          <span class=${`chip ${job.enabled ? "chip-ok" : "chip-danger"}`}>
+            ${job.enabled ? t("cron.jobList.enabled") : t("cron.jobList.disabled")}
+          </span>
+          <span class="chip">${job.sessionTarget}</span>
+          <span class="chip">${job.wakeMode}</span>
+        </div>
+        <div class="row cron-job-actions">
+          <button
+            class="btn"
+            ?disabled=${props.busy}
+            @click=${(event: Event) => {
+              event.stopPropagation();
+              selectAnd(() => props.onEdit(job));
+            }}
+          >
+            ${t("cron.jobList.edit")}
+          </button>
+          <button
+            class="btn"
+            ?disabled=${props.busy}
+            @click=${(event: Event) => {
+              event.stopPropagation();
+              selectAnd(() => props.onClone(job));
+            }}
+          >
+            ${t("cron.jobList.clone")}
+          </button>
+          <button
+            class="btn"
+            ?disabled=${props.busy}
+            @click=${(event: Event) => {
+              event.stopPropagation();
+              selectAnd(() => props.onToggle(job, !job.enabled));
+            }}
+          >
+            ${job.enabled ? t("cron.jobList.disable") : t("cron.jobList.enable")}
+          </button>
+          <button
+            class="btn"
+            ?disabled=${props.busy}
+            @click=${(event: Event) => {
+              event.stopPropagation();
+              selectAnd(() => props.onRun(job, "force"));
+            }}
+          >
+            ${t("cron.jobList.run")}
+          </button>
+          <button
+            class="btn"
+            ?disabled=${props.busy}
+            @click=${(event: Event) => {
+              event.stopPropagation();
+              selectAnd(() => props.onRun(job, "due"));
+            }}
+          >
+            Run if due
+          </button>
+          <button
+            class="btn"
+            ?disabled=${props.busy}
+            @click=${(event: Event) => {
+              event.stopPropagation();
+              selectAnd(() => props.onLoadRuns(job.id));
+            }}
+          >
+            ${t("cron.jobList.history")}
+          </button>
+          <button
+            class="btn danger"
+            ?disabled=${props.busy}
+            @click=${(event: Event) => {
+              event.stopPropagation();
+              selectAnd(() => props.onRemove(job));
+            }}
+          >
+            ${t("cron.jobList.remove")}
+          </button>
+        </div>
+      </div>
+    </div>
+  `;
+}
+
+function renderJobPayload(job: CronJob) {
+  if (job.payload.kind === "systemEvent") {
+    return html`<div class="cron-job-detail">
+      <span class="cron-job-detail-label">${t("cron.jobDetail.system")}</span>
+      <span class="muted cron-job-detail-value">${job.payload.text}</span>
+    </div>`;
+  }
+
+  const delivery = job.delivery;
+  const deliveryTarget =
+    delivery?.mode === "webhook"
+      ? delivery.to
+        ? ` (${delivery.to})`
+        : ""
+      : delivery?.channel || delivery?.to
+        ? ` (${delivery.channel ?? "last"}${delivery.to ? ` -> ${delivery.to}` : ""})`
+        : "";
+
+  return html`
+    <div class="cron-job-detail">
+      <span class="cron-job-detail-label">${t("cron.jobDetail.prompt")}</span>
+      <span class="muted cron-job-detail-value">${job.payload.message}</span>
+    </div>
+    ${
+      delivery
+        ? html`<div class="cron-job-detail">
+            <span class="cron-job-detail-label">${t("cron.jobDetail.delivery")}</span>
+            <span class="muted cron-job-detail-value">${delivery.mode}${deliveryTarget}</span>
+          </div>`
+        : nothing
+    }
+  `;
+}
+
+function renderJobState(job: CronJob) {
+  const rawStatus = job.state?.lastStatus;
+  const statusClass =
+    rawStatus === "ok"
+      ? "cron-job-status-ok"
+      : rawStatus === "error"
+        ? "cron-job-status-error"
+        : rawStatus === "skipped"
+          ? "cron-job-status-skipped"
+          : "cron-job-status-na";
+  const statusLabel =
+    rawStatus === "ok"
+      ? t("cron.runs.runStatusOk")
+      : rawStatus === "error"
+        ? t("cron.runs.runStatusError")
+        : rawStatus === "skipped"
+          ? t("cron.runs.runStatusSkipped")
+          : t("common.na");
+  const nextRunAtMs = job.state?.nextRunAtMs;
+  const lastRunAtMs = job.state?.lastRunAtMs;
+
+  return html`
+    <div class="cron-job-state">
+      <div class="cron-job-state-row">
+        <span class="cron-job-state-key">${t("cron.jobState.status")}</span>
+        <span class=${`cron-job-status-pill ${statusClass}`}>${statusLabel}</span>
+      </div>
+      <div class="cron-job-state-row">
+        <span class="cron-job-state-key">${t("cron.jobState.next")}</span>
+        <span class="cron-job-state-value" title=${formatMs(nextRunAtMs)}>
+          ${formatStateRelative(nextRunAtMs)}
+        </span>
+      </div>
+      <div class="cron-job-state-row">
+        <span class="cron-job-state-key">${t("cron.jobState.last")}</span>
+        <span class="cron-job-state-value" title=${formatMs(lastRunAtMs)}>
+          ${formatStateRelative(lastRunAtMs)}
+        </span>
+      </div>
+    </div>
+  `;
+}
