+++ ui/src/ui/views/cron/runs.ts
@@ -1,0 +1,84 @@
+import { html, nothing } from "lit";
+import { t } from "../../../i18n/index.ts";
+import { formatMs } from "../../format.ts";
+import { pathForTab } from "../../navigation.ts";
+import type { CronRunLogEntry } from "../../types.ts";
+import { formatRunNextLabel } from "./helpers.ts";
+
+function runStatusLabel(value: string): string {
+  switch (value) {
+    case "ok":
+      return t("cron.runs.runStatusOk");
+    case "error":
+      return t("cron.runs.runStatusError");
+    case "skipped":
+      return t("cron.runs.runStatusSkipped");
+    default:
+      return t("cron.runs.runStatusUnknown");
+  }
+}
+
+function runDeliveryLabel(value: string): string {
+  switch (value) {
+    case "delivered":
+      return t("cron.runs.deliveryDelivered");
+    case "not-delivered":
+      return t("cron.runs.deliveryNotDelivered");
+    case "not-requested":
+      return t("cron.runs.deliveryNotRequested");
+    case "unknown":
+      return t("cron.runs.deliveryUnknown");
+    default:
+      return t("cron.runs.deliveryUnknown");
+  }
+}
+
+export function renderRun(entry: CronRunLogEntry, basePath: string) {
+  const chatUrl =
+    typeof entry.sessionKey === "string" && entry.sessionKey.trim().length > 0
+      ? `${pathForTab("chat", basePath)}?session=${encodeURIComponent(entry.sessionKey)}`
+      : null;
+  const status = runStatusLabel(entry.status ?? "unknown");
+  const delivery = runDeliveryLabel(entry.deliveryStatus ?? "not-requested");
+  const usage = entry.usage;
+  const usageSummary =
+    usage && typeof usage.total_tokens === "number"
+      ? `${usage.total_tokens} tokens`
+      : usage && typeof usage.input_tokens === "number" && typeof usage.output_tokens === "number"
+        ? `${usage.input_tokens} in / ${usage.output_tokens} out`
+        : null;
+  return html`
+    <div class="list-item cron-run-entry">
+      <div class="list-main cron-run-entry__main">
+        <div class="list-title cron-run-entry__title">
+          ${entry.jobName ?? entry.jobId}
+          <span class="muted"> · ${status}</span>
+        </div>
+        <div class="list-sub cron-run-entry__summary">${entry.summary ?? entry.error ?? t("cron.runEntry.noSummary")}</div>
+        <div class="chip-row" style="margin-top: 6px;">
+          <span class="chip">${delivery}</span>
+          ${entry.model ? html`<span class="chip">${entry.model}</span>` : nothing}
+          ${entry.provider ? html`<span class="chip">${entry.provider}</span>` : nothing}
+          ${usageSummary ? html`<span class="chip">${usageSummary}</span>` : nothing}
+        </div>
+      </div>
+      <div class="list-meta cron-run-entry__meta">
+        <div>${formatMs(entry.ts)}</div>
+        ${typeof entry.runAtMs === "number" ? html`<div class="muted">${t("cron.runEntry.runAt")} ${formatMs(entry.runAtMs)}</div>` : nothing}
+        <div class="muted">${entry.durationMs ?? 0}ms</div>
+        ${
+          typeof entry.nextRunAtMs === "number"
+            ? html`<div class="muted">${formatRunNextLabel(entry.nextRunAtMs)}</div>`
+            : nothing
+        }
+        ${
+          chatUrl
+            ? html`<div><a class="session-link" href=${chatUrl}>${t("cron.runEntry.openRunChat")}</a></div>`
+            : nothing
+        }
+        ${entry.error ? html`<div class="muted">${entry.error}</div>` : nothing}
+        ${entry.deliveryError ? html`<div class="muted">${entry.deliveryError}</div>` : nothing}
+      </div>
+    </div>
+  `;
+}
